+++ app/build.gradle
@@ -75,9 +75,6 @@ dependencies {
 
     implementation "com.jakewharton.timber:timber:4.7.1"
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:1.5.0"
-    implementation "com.fasterxml.jackson.core:jackson-core:2.13.2"
-    implementation "com.bluelinelabs:logansquare:1.3.7"
-    kapt "com.bluelinelabs:logansquare-compiler:1.3.7"
 
     testImplementation 'junit:junit:4.13.2'
 }
