+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     <application
         android:name=".Game"
         android:icon="@drawable/ic_launcher"
-        android:label="@string/app_name">
+        android:label="Tomb">
         <activity
             android:name=".MainActivity"
             android:configChanges="keyboardHidden|orientation"
