+++ app/src/main/java/ru/meatgames/tomb/domain/render/GameMapRenderPipeline.kt
@@ -6,13 +6,13 @@ import ru.meatgames.tomb.domain.Coordinates
 import ru.meatgames.tomb.domain.enemy.EnemiesHolder
 import ru.meatgames.tomb.domain.item.ItemsHolder
 import ru.meatgames.tomb.domain.ScreenSpaceCoordinates
-import ru.meatgames.tomb.model.temp.ThemeAssets
+import ru.meatgames.tomb.model.theme.ThemeAssets
 import ru.meatgames.tomb.model.tile.domain.FloorRenderTile
 import ru.meatgames.tomb.model.tile.domain.ObjectRenderTile
 import ru.meatgames.tomb.render.MapRenderTile
 import ru.meatgames.tomb.render.RenderData
 import ru.meatgames.tomb.domain.map.MapTileWrapper
-import ru.meatgames.tomb.model.temp.ASSETS_TILE_SIZE
+import ru.meatgames.tomb.model.theme.ASSETS_TILE_SIZE
 import javax.inject.Inject
 
 class GameMapRenderPipeline @Inject constructor(
