+++ app/src/main/java/ru/meatgames/tomb/domain/render/RoomPreviewRenderProcessor.kt
@@ -3,8 +3,8 @@ package ru.meatgames.tomb.domain.render
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.IntOffset
 import ru.meatgames.tomb.domain.map.MapTileWrapper
-import ru.meatgames.tomb.model.temp.ASSETS_TILE_SIZE
-import ru.meatgames.tomb.model.temp.ThemeAssets
+import ru.meatgames.tomb.model.theme.ASSETS_TILE_SIZE
+import ru.meatgames.tomb.model.theme.ThemeAssets
 import ru.meatgames.tomb.model.tile.domain.FloorEntityTile
 import ru.meatgames.tomb.model.tile.domain.FloorRenderTile
 import ru.meatgames.tomb.model.tile.domain.ObjectEntityTile
