+++ app/src/main/java/ru/meatgames/tomb/model/theme/FloorThemes.kt
@@ -1,17 +1,17 @@
-package ru.meatgames.tomb.model.temp
+package ru.meatgames.tomb.model.theme
 
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.IntOffset
-import ru.meatgames.tomb.model.tile.domain.ObjectRenderTile
+import ru.meatgames.tomb.model.tile.domain.FloorRenderTile
 
-data class DoorsThemes(
+data class FloorThemes(
     val atlas: ImageBitmap,
     val themes: List<Theme>,
 ) {
 
     data class Theme(
         val title: String,
-        val map: Map<ObjectRenderTile, IntOffset>,
+        val map: Map<FloorRenderTile, IntOffset>,
     )
 
 }
