+++ app/src/main/java/ru/meatgames/tomb/model/theme/StairTilesDto.kt
@@ -1,4 +1,4 @@
-package ru.meatgames.tomb.model.temp
+package ru.meatgames.tomb.model.theme
 
 import kotlinx.serialization.Serializable
 import ru.meatgames.tomb.model.tile.domain.ObjectRenderTile
