+++ app/src/main/java/ru/meatgames/tomb/model/theme/StairsThemes.kt
@@ -1,17 +1,17 @@
-package ru.meatgames.tomb.model.temp
+package ru.meatgames.tomb.model.theme
 
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.IntOffset
-import ru.meatgames.tomb.model.tile.domain.FloorRenderTile
+import ru.meatgames.tomb.model.tile.domain.ObjectRenderTile
 
-data class FloorThemes(
+data class StairsThemes(
     val atlas: ImageBitmap,
     val themes: List<Theme>,
 ) {
 
     data class Theme(
         val title: String,
-        val map: Map<FloorRenderTile, IntOffset>,
+        val map: Map<ObjectRenderTile, IntOffset>,
     )
 
 }
