+++ app/src/main/java/ru/meatgames/tomb/model/theme/WallsThemes.kt
@@ -1,4 +1,4 @@
-package ru.meatgames.tomb.model.temp
+package ru.meatgames.tomb.model.theme
 
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.IntOffset
