+++ app/src/main/java/ru/meatgames/tomb/screen/compose/RoomRenderer.kt
@@ -23,13 +23,13 @@ import androidx.compose.ui.unit.dp
 import ru.meatgames.tomb.design.h2TextStyle
 import ru.meatgames.tomb.domain.render.RoomPreviewRenderProcessor
 import ru.meatgames.tomb.model.room.data.RoomsRepository
-import ru.meatgames.tomb.model.temp.ThemeAssets
+import ru.meatgames.tomb.model.theme.ThemeAssets
 import ru.meatgames.tomb.render.MapRenderTile
 import ru.meatgames.tomb.render.WallsDecorator
 import ru.meatgames.tomb.domain.map.MapTile
 import ru.meatgames.tomb.domain.map.MapTileWrapper
-import ru.meatgames.tomb.model.temp.ASSETS_TILE_DIMENSION
-import ru.meatgames.tomb.model.temp.ASSETS_TILE_SIZE
+import ru.meatgames.tomb.model.theme.ASSETS_TILE_DIMENSION
+import ru.meatgames.tomb.model.theme.ASSETS_TILE_SIZE
 import kotlin.math.max
 
 @Preview
