+++ app/src/main/java/ru/meatgames/tomb/screen/compose/game/component/GameScreenPreviewData.kt
@@ -5,8 +5,8 @@ import ru.meatgames.tomb.domain.enemy.EnemyType
 import ru.meatgames.tomb.domain.enemy.produceEnemy
 import ru.meatgames.tomb.domain.map.MapScreenState
 import ru.meatgames.tomb.domain.render.hasBottomShadow
-import ru.meatgames.tomb.model.temp.ASSETS_TILE_SIZE
-import ru.meatgames.tomb.model.temp.ThemeAssets
+import ru.meatgames.tomb.model.theme.ASSETS_TILE_SIZE
+import ru.meatgames.tomb.model.theme.ThemeAssets
 import ru.meatgames.tomb.model.tile.domain.FloorRenderTile
 import ru.meatgames.tomb.model.tile.domain.ObjectRenderTile
 import ru.meatgames.tomb.render.MapRenderTile
